Update: init can receive authors, parents, follows or contents
