+++ lib/utils.js
@@ -60,3 +60,40 @@ const JSONdecode = (exports.JSONdecode = str => {
   }
   return _JSONdecode(str, transform)
 })
+
+// helper dat.json object mould
+exports.createDatJSON = ({
+  title,
+  description,
+  url,
+  license = [
+    {
+      href: 'https://creativecommons.org/publicdomain/zero/1.0/legalcode'
+    }
+  ],
+  spec = [{ href: 'https://p2pcommons.com/specs/module/0.2.0' }],
+  authors,
+  follows,
+  contents,
+  parents,
+  ...p2pcommons
+}) => {
+  const mould = {
+    title,
+    description,
+    url,
+    links: { license, spec },
+    p2pcommons
+  }
+  // TODO(dk): validate links (license, spec)
+
+  if (mould.p2pcommons.type === 'profile') {
+    mould.p2pcommons.follows = follows || []
+    mould.p2pcommons.contents = contents || []
+  } else {
+    mould.p2pcommons.authors = authors || []
+    mould.p2pcommons.parents = parents || []
+  }
+
+  return mould
+}
